+++ winui3/SquirrelNotifier.WinUI3/Services/IReviewLauncherService.cs
@@ -1,0 +1,18 @@
+// <copyright file="IReviewLauncherService.cs" company="PlaceholderCompany">
+// Copyright (c) PlaceholderCompany. All rights reserved.
+// </copyright>
+
+using System.Threading;
+using System.Threading.Tasks;
+using SquirrelNotifier.WinUI3.Models;
+
+namespace SquirrelNotifier.WinUI3.Services;
+
+internal interface IReviewLauncherService
+{
+    bool IsRunning { get; }
+
+    Task<LauncherResult> LaunchAsync(ReviewEvent reviewEvent, CancellationToken cancellationToken);
+
+    void Cancel();
+}